Fix compilation warnings

+++ src/robotAbstract.cpp
@@ -154,18 +154,18 @@ namespace jafar {
 			std::ostringstream oss; oss << "Robot " << id();
 			log.writeComment(oss.str());
 			log.writeLegendTokens("time");
-			for(int i = 0; i < state.x().size(); ++i)
+			for(size_t i = 0; i < state.x().size(); ++i)
 				{ oss.str(""); oss << "x" << i; log.writeLegend(oss.str()); }
-			for(int i = 0; i < state.x().size(); ++i)
+			for(size_t i = 0; i < state.x().size(); ++i)
 				{ oss.str(""); oss << "sig" << i; log.writeLegend(oss.str()); }
 		}
 		
 		void RobotAbstract::writeLogData(kernel::DataLogger& log) const
 		{
 			log.writeData(self_time);
-			for(int i = 0; i < state.x().size(); ++i)
+			for(size_t i = 0; i < state.x().size(); ++i)
 				log.writeData(state.x()(i));
-			for(int i = 0; i < state.x().size(); ++i)
+			for(size_t i = 0; i < state.x().size(); ++i)
 				log.writeData(sqrt(state.P()(i,i)));
 		}
